requirements.txt
google-generativeai
huggingface_hub
supabase
langchain-google-genai
langchain